+++ ddtrace/contrib/flask/__init__.py
@@ -36,7 +36,11 @@ required_modules = ['flask']
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
+        # DEV: We do this so we can `@mock.patch('ddtrace.contrib.flask._patch.<func>')` in tests
+        from . import patch as _patch
         from .middleware import TraceMiddleware
-        from .monkey import patch, unpatch
+
+        patch = _patch.patch
+        unpatch = _patch.unpatch
 
         __all__ = ['TraceMiddleware', 'patch', 'unpatch']
