+++ tests/contrib/flask/test_idempotency.py
@@ -5,7 +5,7 @@ import flask
 import wrapt
 
 from ddtrace.contrib.flask import patch, unpatch
-from ddtrace.contrib.flask.monkey import _w, _u
+from ddtrace.contrib.flask.patch import _w, _u
 
 
 class FlaskIdempotencyTestCase(unittest.TestCase):
@@ -36,7 +36,7 @@ class FlaskIdempotencyTestCase(unittest.TestCase):
         self.assert_is_not_patched()
 
     # DEV: Use `side_effect` so the original function still gets called
-    @mock.patch('ddtrace.contrib.flask.monkey._w', side_effect=_w)
+    @mock.patch('ddtrace.contrib.flask._patch._w', side_effect=_w)
     def test_patch_idempotency(self, _w):
         # Ensure we didn't do any patching automatically
         _w.assert_not_called()
@@ -56,8 +56,8 @@ class FlaskIdempotencyTestCase(unittest.TestCase):
         self.assert_is_patched()
 
     # DEV: Use `side_effect` so the original function still gets called
-    @mock.patch('ddtrace.contrib.flask.monkey._w', side_effect=_w)
-    @mock.patch('ddtrace.contrib.flask.monkey._u', side_effect=_u)
+    @mock.patch('ddtrace.contrib.flask._patch._w', side_effect=_w)
+    @mock.patch('ddtrace.contrib.flask._patch._u', side_effect=_u)
     def test_unpatch_idempotency(self, _u, _w):
         # We need to patch in order to unpatch
         patch()
